Moved function 'resetBooking' out of MovieBookingComponent
